+++ examples/simple/simpleUpload.js
@@ -101,13 +101,13 @@ var http = require( 'http' ),
                     // default is false
                     // it serves to monitor the progress of the file upload
                     // and also to move the data to another stream, while the file is being uploaded 
-                emitFileProgress : !false,
+                emitFileProgress : false,
                     
                     // switch on/off 'progress' event
                     // default is false, or integer chunk factor, 
                     // every n chunk emits a dataprogress event:  1 + ( 0 * n ) 1 + ( 1 * n ), 1 + ( 2 * n ), 1 + ( 3 * n ), 
                     // minimum factor value is 2 
-                emitProgress : !false, // 3, 10, 100
+                emitProgress : false, // 3, 10, 100
                 
                     // max bytes allowed for file uploads ( multipart/form-data ), it is a writing threshold, this is the max size of bytes written to disk before stopping
                 uploadThreshold : 1024 * 1024 * 1024 ,// bytes
@@ -137,7 +137,7 @@ var http = require( 'http' ),
                     // file: 'on' --> create a log file in the current upload directory with the same name and .log extension
                     // console: 'off' --> disable console log output 
                     // record: 'on' --> record binary data from client request
-                logging : 'debug:on,1:on,2:on,3:on,4:on,console:on,file:off,record:off', // <-- turn off 2nd level to see only warnings, and parser overall results
+                logging : 'debug:on,1:on,2:on,3:off,4:off,console:on,file:off,record:off', // <-- turn off 2nd level to see only warnings, and parser overall results
                 
                     // listeners
                 listeners : {